stock_forecast/requirements_deploy.txt
streamlit==1.31.1
pandas==2.1.3
plotly==5.18.0
numpy==1.24.3
scikit-learn==1.3.2
python-dotenv==1.0.0
supabase==2.4.0
